requirements.txt
livekit-agents>=1.4.5
livekit-api>=1.1.0
livekit-plugins-openai>=1.4.5
livekit-plugins-cartesia>=1.4.5
livekit-plugins-deepgram>=1.4.5
livekit-plugins-silero>=1.4.5
livekit-plugins-noise-cancellation>=0.2.5
livekit-plugins-turn-detector>=1.4.5
livekit-plugins-sarvam>=1.4.5
python-dotenv>=1.0.0
supabase
